Add konfig resource to sync configmaps
This commit introduces a CRD konfig in group krius.dev and
version v1alpha1 that can be used to create configmaps
across all the namespaes.

if we just run

```
k create -f examples/manifests/konfig.yaml
```

configmap in all the namespace would be created with data
that is provided in the konfig cr.